Introduce RELEASE_EXTERNAL_TC3_DEBUG flag
It's used to enable assertions and change log verbosity in
libtextclassifier package, which are currently controlled using the
debuggable know in the package blueprint.
The value is set to false on user and true on userdebug/eng builds,
maintaining the existing behavior. Therefore this is intended to be
a non-functional change.

BUG: b/373580227
Test: lunch tangorpro-trunk_staging-{user/userdebug/eng} and check the
      contents of out/soong/release-config/build_flags_product.json
Change-Id: I46df139b15694c49dbeb1698dc6c7a00d0342bc7