Add MediaProvider TEST_MAPPING to trunk_staging config
This will ensure that MediaProvider presubmits are executed
when we try to rollout flag to staging. This will catch
any presubmit failures and stop flag rollout. Presently
existing mts/cts tests are not executed with new code
changes enabled till the flag is completely rolled out
to trunk staging and if any error is encountered a bug
is created and flag is rolled back.

Test: n/a
Bug: 401230775
Flag: EXEMPT TEST_MAPPING addition
Ignore-AOSP-First: internal process
Change-Id: Id098a4efcf94517c15d61690e38834e2ed58816c