create build flag RELEASE_EXPORTED_FLAG_CHECK
The flag is to control the exported flag check when the build system
build all_aconfig_declarations. Once this flag is true, when the build
system builds all_aconfig_declarations, it will check whether all
exported flags are used for guarding APIs. The build will fail if
existing on exported flag that is not used for guarding any API.

Test: set the flag as true, and build locally
Bug: 336800305
Change-Id: Iaeb8551615221278271992cd6edcdff27cadbf02